+++ crates/pgt_completions/Cargo.toml
@@ -19,7 +19,7 @@ pgt_text_size.workspace = true
 
 fuzzy-matcher                    = "0.3.7"
 pgt_schema_cache.workspace       = true
-pgt_treesitter_queries.workspace = true
+pgt_treesitter.workspace = true
 schemars                         = { workspace = true, optional = true }
 serde                            = { workspace = true, features = ["derive"] }
 serde_json                       = { workspace = true }
