+++ crates/pgt_completions/src/providers/columns.rs
@@ -49,10 +49,12 @@ mod tests {
     use crate::{
         CompletionItem, CompletionItemKind, complete,
         test_helper::{
-            CURSOR_POS, CompletionAssertion, InputQuery, assert_complete_results,
-            assert_no_complete_results, get_test_deps, get_test_params,
+            CompletionAssertion, assert_complete_results, assert_no_complete_results,
+            get_test_deps, get_test_params,
         },
     };
+
+    use pgt_test_utils::QueryWithCursorPosition;
 
     struct TestCase {
         query: String,
@@ -62,7 +64,7 @@ mod tests {
     }
 
     impl TestCase {
-        fn get_input_query(&self) -> InputQuery {
+        fn get_input_query(&self) -> QueryWithCursorPosition {
             let strs: Vec<&str> = self.query.split_whitespace().collect();
             strs.join(" ").as_str().into()
         }
@@ -94,7 +96,10 @@ mod tests {
         let queries: Vec<TestCase> = vec![
             TestCase {
                 message: "correctly prefers the columns of present tables",
-                query: format!(r#"select na{} from public.audio_books;"#, CURSOR_POS),
+                query: format!(
+                    r#"select na{} from public.audio_books;"#,
+                    QueryWithCursorPosition::cursor_marker()
+                ),
                 label: "narrator",
                 description: "public.audio_books",
             },
@@ -111,14 +116,17 @@ mod tests {
                 join public.users u
                 on u.id = subquery.id;
                 "#,
-                    CURSOR_POS
+                    QueryWithCursorPosition::cursor_marker()
                 ),
                 label: "narrator_id",
                 description: "private.audio_books",
             },
             TestCase {
                 message: "works without a schema",
-                query: format!(r#"select na{} from users;"#, CURSOR_POS),
+                query: format!(
+                    r#"select na{} from users;"#,
+                    QueryWithCursorPosition::cursor_marker()
+                ),
                 label: "name",
                 description: "public.users",
             },
@@ -165,7 +173,7 @@ mod tests {
         pool.execute(setup).await.unwrap();
 
         let case = TestCase {
-            query: format!(r#"select n{};"#, CURSOR_POS),
+            query: format!(r#"select n{};"#, QueryWithCursorPosition::cursor_marker()),
             description: "",
             label: "",
             message: "",
@@ -220,7 +228,10 @@ mod tests {
 
         let test_case = TestCase {
             message: "suggests user created tables first",
-            query: format!(r#"select {} from users"#, CURSOR_POS),
+            query: format!(
+                r#"select {} from users"#,
+                QueryWithCursorPosition::cursor_marker()
+            ),
             label: "",
             description: "",
         };
@@ -270,7 +281,10 @@ mod tests {
 
         let test_case = TestCase {
             message: "suggests user created tables first",
-            query: format!(r#"select * from private.{}"#, CURSOR_POS),
+            query: format!(
+                r#"select * from private.{}"#,
+                QueryWithCursorPosition::cursor_marker()
+            ),
             label: "",
             description: "",
         };
@@ -311,7 +325,11 @@ mod tests {
         pool.execute(setup).await.unwrap();
 
         assert_complete_results(
-            format!(r#"select {} from users"#, CURSOR_POS).as_str(),
+            format!(
+                r#"select {} from users"#,
+                QueryWithCursorPosition::cursor_marker()
+            )
+            .as_str(),
             vec![
                 CompletionAssertion::Label("address2".into()),
                 CompletionAssertion::Label("email2".into()),
@@ -324,7 +342,11 @@ mod tests {
         .await;
 
         assert_complete_results(
-            format!(r#"select {} from private.users"#, CURSOR_POS).as_str(),
+            format!(
+                r#"select {} from private.users"#,
+                QueryWithCursorPosition::cursor_marker()
+            )
+            .as_str(),
             vec![
                 CompletionAssertion::Label("address1".into()),
                 CompletionAssertion::Label("email1".into()),
@@ -338,7 +360,11 @@ mod tests {
 
         // asserts fuzzy finding for "settings"
         assert_complete_results(
-            format!(r#"select sett{} from private.users"#, CURSOR_POS).as_str(),
+            format!(
+                r#"select sett{} from private.users"#,
+                QueryWithCursorPosition::cursor_marker()
+            )
+            .as_str(),
             vec![CompletionAssertion::Label("user_settings".into())],
             None,
             &pool,
@@ -372,7 +398,7 @@ mod tests {
         assert_complete_results(
             format!(
                 "select u.id, p.{} from auth.users u join auth.posts p on u.id = p.user_id;",
-                CURSOR_POS
+                QueryWithCursorPosition::cursor_marker()
             )
             .as_str(),
             vec![
@@ -394,7 +420,7 @@ mod tests {
         assert_complete_results(
             format!(
                 "select u.id, p.content from auth.users u join auth.posts p on u.id = p.{};",
-                CURSOR_POS
+                QueryWithCursorPosition::cursor_marker()
             )
             .as_str(),
             vec![
@@ -440,7 +466,7 @@ mod tests {
         assert_complete_results(
             format!(
                 "select u.id, p.content from auth.users u join auth.{}",
-                CURSOR_POS
+                QueryWithCursorPosition::cursor_marker()
             )
             .as_str(),
             vec![
@@ -479,7 +505,7 @@ mod tests {
         assert_complete_results(
             format!(
                 "select u.id, auth.posts.content from auth.users u join auth.posts on u.{}",
-                CURSOR_POS
+                QueryWithCursorPosition::cursor_marker()
             )
             .as_str(),
             vec![
@@ -496,7 +522,7 @@ mod tests {
         assert_complete_results(
             format!(
                 "select u.id, p.content from auth.users u join auth.posts p on p.user_id = u.{}",
-                CURSOR_POS
+                QueryWithCursorPosition::cursor_marker()
             )
             .as_str(),
             vec![
@@ -536,7 +562,7 @@ mod tests {
         assert_complete_results(
             format!(
                 "select {} from public.one o join public.two on o.id = t.id;",
-                CURSOR_POS
+                QueryWithCursorPosition::cursor_marker()
             )
             .as_str(),
             vec![
@@ -555,7 +581,7 @@ mod tests {
         assert_complete_results(
             format!(
                 "select a, {} from public.one o join public.two on o.id = t.id;",
-                CURSOR_POS
+                QueryWithCursorPosition::cursor_marker()
             )
             .as_str(),
             vec![
@@ -577,7 +603,7 @@ mod tests {
         assert_complete_results(
             format!(
                 "select o.id, a, b, c, d, e, {} from public.one o join public.two on o.id = t.id;",
-                CURSOR_POS
+                QueryWithCursorPosition::cursor_marker()
             )
             .as_str(),
             vec![
@@ -593,7 +619,7 @@ mod tests {
         assert_complete_results(
             format!(
                 "select id, a, b, c, d, e, {} from public.one o join public.two on o.id = t.id;",
-                CURSOR_POS
+                QueryWithCursorPosition::cursor_marker()
             )
             .as_str(),
             vec![CompletionAssertion::Label("z".to_string())],
@@ -625,7 +651,11 @@ mod tests {
         // are lower in the alphabet
 
         assert_complete_results(
-            format!("insert into instruments ({})", CURSOR_POS).as_str(),
+            format!(
+                "insert into instruments ({})",
+                QueryWithCursorPosition::cursor_marker()
+            )
+            .as_str(),
             vec![
                 CompletionAssertion::Label("id".to_string()),
                 CompletionAssertion::Label("name".to_string()),
@@ -637,7 +667,11 @@ mod tests {
         .await;
 
         assert_complete_results(
-            format!("insert into instruments (id, {})", CURSOR_POS).as_str(),
+            format!(
+                "insert into instruments (id, {})",
+                QueryWithCursorPosition::cursor_marker()
+            )
+            .as_str(),
             vec![
                 CompletionAssertion::Label("name".to_string()),
                 CompletionAssertion::Label("z".to_string()),
@@ -648,7 +682,11 @@ mod tests {
         .await;
 
         assert_complete_results(
-            format!("insert into instruments (id, {}, name)", CURSOR_POS).as_str(),
+            format!(
+                "insert into instruments (id, {}, name)",
+                QueryWithCursorPosition::cursor_marker()
+            )
+            .as_str(),
             vec![CompletionAssertion::Label("z".to_string())],
             None,
             &pool,
@@ -659,7 +697,7 @@ mod tests {
         assert_complete_results(
             format!(
                 "insert into instruments (name, {}) values ('my_bass');",
-                CURSOR_POS
+                QueryWithCursorPosition::cursor_marker()
             )
             .as_str(),
             vec![
@@ -673,7 +711,11 @@ mod tests {
 
         // no completions in the values list!
         assert_no_complete_results(
-            format!("insert into instruments (id, name) values ({})", CURSOR_POS).as_str(),
+            format!(
+                "insert into instruments (id, name) values ({})",
+                QueryWithCursorPosition::cursor_marker()
+            )
+            .as_str(),
             None,
             &pool,
         )
@@ -700,7 +742,11 @@ mod tests {
         pool.execute(setup).await.unwrap();
 
         assert_complete_results(
-            format!("select name from instruments where {} ", CURSOR_POS).as_str(),
+            format!(
+                "select name from instruments where {} ",
+                QueryWithCursorPosition::cursor_marker()
+            )
+            .as_str(),
             vec![
                 CompletionAssertion::Label("created_at".into()),
                 CompletionAssertion::Label("id".into()),
@@ -715,7 +761,7 @@ mod tests {
         assert_complete_results(
             format!(
                 "select name from instruments where z = 'something' and created_at > {}",
-                CURSOR_POS
+                QueryWithCursorPosition::cursor_marker()
             )
             .as_str(),
             // simply do not complete columns + schemas; functions etc. are ok
@@ -732,7 +778,7 @@ mod tests {
         assert_complete_results(
             format!(
                 "select name from instruments where id = 'something' and {}",
-                CURSOR_POS
+                QueryWithCursorPosition::cursor_marker()
             )
             .as_str(),
             vec![
@@ -749,7 +795,7 @@ mod tests {
         assert_complete_results(
             format!(
                 "select name from instruments i join others o on i.z = o.a where i.{}",
-                CURSOR_POS
+                QueryWithCursorPosition::cursor_marker()
             )
             .as_str(),
             vec![
@@ -783,22 +829,37 @@ mod tests {
         pool.execute(setup).await.unwrap();
 
         let queries = vec![
-            format!("alter table instruments drop column {}", CURSOR_POS),
+            format!(
+                "alter table instruments drop column {}",
+                QueryWithCursorPosition::cursor_marker()
+            ),
             format!(
                 "alter table instruments drop column if exists {}",
-                CURSOR_POS
+                QueryWithCursorPosition::cursor_marker()
             ),
             format!(
                 "alter table instruments alter column {} set default",
-                CURSOR_POS
+                QueryWithCursorPosition::cursor_marker()
             ),
-            format!("alter table instruments alter {} set default", CURSOR_POS),
-            format!("alter table public.instruments alter column {}", CURSOR_POS),
-            format!("alter table instruments alter {}", CURSOR_POS),
-            format!("alter table instruments rename {} to new_col", CURSOR_POS),
+            format!(
+                "alter table instruments alter {} set default",
+                QueryWithCursorPosition::cursor_marker()
+            ),
+            format!(
+                "alter table public.instruments alter column {}",
+                QueryWithCursorPosition::cursor_marker()
+            ),
+            format!(
+                "alter table instruments alter {}",
+                QueryWithCursorPosition::cursor_marker()
+            ),
+            format!(
+                "alter table instruments rename {} to new_col",
+                QueryWithCursorPosition::cursor_marker()
+            ),
             format!(
                 "alter table public.instruments rename column {} to new_col",
-                CURSOR_POS
+                QueryWithCursorPosition::cursor_marker()
             ),
         ];
 
@@ -834,19 +895,19 @@ mod tests {
         let col_queries = vec![
             format!(
                 r#"create policy "my_pol" on public.instruments for select using ({})"#,
-                CURSOR_POS
+                QueryWithCursorPosition::cursor_marker()
             ),
             format!(
                 r#"create policy "my_pol" on public.instruments for insert with check ({})"#,
-                CURSOR_POS
+                QueryWithCursorPosition::cursor_marker()
             ),
             format!(
                 r#"create policy "my_pol" on public.instruments for update using (id = 1 and {})"#,
-                CURSOR_POS
+                QueryWithCursorPosition::cursor_marker()
             ),
             format!(
                 r#"create policy "my_pol" on public.instruments for insert with check (id = 1 and {})"#,
-                CURSOR_POS
+                QueryWithCursorPosition::cursor_marker()
             ),
         ];
