+++ crates/pgt_completions/src/test_helper.rs
@@ -1,40 +1,14 @@
 use std::fmt::Display;
 
 use pgt_schema_cache::SchemaCache;
+use pgt_test_utils::QueryWithCursorPosition;
 use sqlx::{Executor, PgPool};
 
 use crate::{CompletionItem, CompletionItemKind, CompletionParams, complete};
 
-pub static CURSOR_POS: char = '€';
-
-#[derive(Clone)]
-pub struct InputQuery {
-    sql: String,
-    position: usize,
-}
-
-impl From<&str> for InputQuery {
-    fn from(value: &str) -> Self {
-        let position = value
-            .find(CURSOR_POS)
-            .expect("Insert Cursor Position into your Query.");
-
-        InputQuery {
-            sql: value.replace(CURSOR_POS, "").trim().to_string(),
-            position,
-        }
-    }
-}
-
-impl Display for InputQuery {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.sql)
-    }
-}
-
 pub(crate) async fn get_test_deps(
     setup: Option<&str>,
-    input: InputQuery,
+    input: QueryWithCursorPosition,
     test_db: &PgPool,
 ) -> (tree_sitter::Tree, pgt_schema_cache::SchemaCache) {
     if let Some(setup) = setup {
@@ -63,7 +37,7 @@ pub(crate) async fn get_test_deps(
 #[allow(dead_code)]
 pub(crate) async fn test_against_connection_string(
     conn_str: &str,
-    input: InputQuery,
+    input: QueryWithCursorPosition,
 ) -> (tree_sitter::Tree, pgt_schema_cache::SchemaCache) {
     let pool = sqlx::PgPool::connect(conn_str)
         .await
@@ -83,14 +57,10 @@ pub(crate) async fn test_against_connection_string(
     (tree, schema_cache)
 }
 
-pub(crate) fn get_text_and_position(q: InputQuery) -> (usize, String) {
-    (q.position, q.sql)
-}
-
 pub(crate) fn get_test_params<'a>(
     tree: &'a tree_sitter::Tree,
     schema_cache: &'a pgt_schema_cache::SchemaCache,
-    sql: InputQuery,
+    sql: QueryWithCursorPosition,
 ) -> CompletionParams<'a> {
     let (position, text) = get_text_and_position(sql);
 
@@ -99,46 +69,6 @@ pub(crate) fn get_test_params<'a>(
         schema: schema_cache,
         tree,
         text,
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::test_helper::CURSOR_POS;
-
-    use super::InputQuery;
-
-    #[test]
-    fn input_query_should_extract_correct_position() {
-        struct TestCase {
-            query: String,
-            expected_pos: usize,
-            expected_sql_len: usize,
-        }
-
-        let cases = vec![
-            TestCase {
-                query: format!("select * from{}", CURSOR_POS),
-                expected_pos: 13,
-                expected_sql_len: 13,
-            },
-            TestCase {
-                query: format!("{}select * from", CURSOR_POS),
-                expected_pos: 0,
-                expected_sql_len: 13,
-            },
-            TestCase {
-                query: format!("select {} from", CURSOR_POS),
-                expected_pos: 7,
-                expected_sql_len: 12,
-            },
-        ];
-
-        for case in cases {
-            let query = InputQuery::from(case.query.as_str());
-            assert_eq!(query.position, case.expected_pos);
-            assert_eq!(query.sql.len(), case.expected_sql_len);
-        }
     }
 }
