+++ crates/pgt_treesitter/Cargo.toml
@@ -6,7 +6,7 @@ edition.workspace    = true
 homepage.workspace   = true
 keywords.workspace   = true
 license.workspace    = true
-name                 = "pgt_treesitter_queries"
+name                 = "pgt_treesitter"
 repository.workspace = true
 version              = "0.0.0"
 
@@ -15,6 +15,8 @@ version              = "0.0.0"
 clap                      = { version = "4.5.23", features = ["derive"] }
 tree-sitter.workspace     = true
 tree_sitter_sql.workspace = true
+pgt_text_size.workspace = true
+pgt_schema_cache.workspace = true
 
 [dev-dependencies]
