+++ crates/pgt_treesitter/src/context/grant_parser.rs
@@ -5,7 +5,7 @@ use crate::context::base_parser::{
 };
 
 #[derive(Default, Debug, PartialEq, Eq)]
-pub(crate) struct RevokeContext {
+pub(crate) struct GrantContext {
     pub table_name: Option<String>,
     pub schema_name: Option<String>,
     pub node_text: String,
@@ -17,22 +17,21 @@ pub(crate) struct RevokeContext {
 /// completions.
 /// The parser will only work if the (trimmed) sql starts with `create policy`, `drop policy`, or `alter policy`.
 /// It can only parse policy statements.
-pub(crate) struct RevokeParser {
+pub(crate) struct GrantParser {
     navigator: TokenNavigator,
-    context: RevokeContext,
+    context: GrantContext,
     cursor_position: usize,
     in_roles_list: bool,
-    is_revoking_role: bool,
 }
 
-impl CompletionStatementParser for RevokeParser {
-    type Context = RevokeContext;
-    const NAME: &'static str = "RevokeParser";
+impl CompletionStatementParser for GrantParser {
+    type Context = GrantContext;
+    const NAME: &'static str = "GrantParser";
 
     fn looks_like_matching_stmt(sql: &str) -> bool {
         let lowercased = sql.to_ascii_lowercase();
         let trimmed = lowercased.trim();
-        trimmed.starts_with("revoke")
+        trimmed.starts_with("grant")
     }
 
     fn parse(mut self) -> Self::Context {
@@ -50,15 +49,14 @@ impl CompletionStatementParser for RevokeParser {
     fn make_parser(tokens: Vec<WordWithIndex>, cursor_position: usize) -> Self {
         Self {
             navigator: tokens.into(),
-            context: RevokeContext::default(),
+            context: GrantContext::default(),
             cursor_position,
             in_roles_list: false,
-            is_revoking_role: false,
         }
     }
 }
 
-impl RevokeParser {
+impl GrantParser {
     fn handle_token_under_cursor(&mut self, token: WordWithIndex) {
         if self.navigator.previous_token.is_none() {
             return;
@@ -76,27 +74,24 @@ impl RevokeParser {
             .to_ascii_lowercase()
             .as_str()
         {
+            "grant" => {
+                self.context.node_range = token.get_range();
+                self.context.node_kind = "grant_role".into();
+                self.context.node_text = token.get_word();
+            }
             "on" if !matches!(current.as_deref(), Some("table")) => self.handle_table(&token),
 
             "table" => {
                 self.handle_table(&token);
             }
-
-            "from" | "revoke" => {
+            "to" => {
                 self.context.node_range = token.get_range();
-                self.context.node_kind = "revoke_role".into();
+                self.context.node_kind = "grant_role".into();
                 self.context.node_text = token.get_word();
             }
-
-            "for" if self.is_revoking_role => {
-                self.context.node_range = token.get_range();
-                self.context.node_kind = "revoke_role".into();
-                self.context.node_text = token.get_word();
-            }
-
             t => {
                 if self.in_roles_list && t.ends_with(',') {
-                    self.context.node_kind = "revoke_role".into();
+                    self.context.node_kind = "grant_role".into();
                 }
 
                 self.context.node_range = token.get_range();
@@ -121,7 +116,7 @@ impl RevokeParser {
                 .expect("Text too long");
 
             self.context.node_range = range_without_schema;
-            self.context.node_kind = "revoke_table".into();
+            self.context.node_kind = "grant_table".into();
 
             // In practice, we should always have a table name.
             // The completion sanitization will add a word after a `.` if nothing follows it;
@@ -130,23 +125,35 @@ impl RevokeParser {
         } else {
             self.context.node_range = token.get_range();
             self.context.node_text = token.get_word();
-            self.context.node_kind = "revoke_table".into();
+            self.context.node_kind = "grant_table".into();
         }
     }
 
     fn handle_token(&mut self, token: WordWithIndex) {
         match token.get_word_without_quotes().as_str() {
-            "on" if !self.navigator.next_matches(&["table"]) => self.table_with_schema(),
-
-            // This is the only case where there is no "GRANT" before the option:
-            // REVOKE [ { ADMIN | INHERIT | SET } OPTION FOR ] role_name
-            "option" if !self.navigator.prev_matches(&["grant"]) => {
-                self.is_revoking_role = true;
+            "on" if !self.navigator.next_matches(&[
+                "table",
+                "schema",
+                "foreign",
+                "domain",
+                "sequence",
+                "database",
+                "function",
+                "procedure",
+                "routine",
+                "language",
+                "large",
+                "parameter",
+                "schema",
+                "tablespace",
+                "type",
+            ]) =>
+            {
+                self.table_with_schema()
             }
-
             "table" => self.table_with_schema(),
 
-            "from" => {
+            "to" => {
                 self.in_roles_list = true;
             }
 
@@ -179,9 +186,10 @@ mod tests {
 
     use crate::{
         context::base_parser::CompletionStatementParser,
-        context::revoke_parser::{RevokeContext, RevokeParser},
-        test_helper::CURSOR_POS,
+        context::grant_parser::{GrantContext, GrantParser},
     };
+
+    static CURSOR_POS: char = '€';
 
     fn with_pos(query: String) -> (usize, String) {
         let mut pos: Option<usize> = None;
@@ -200,24 +208,24 @@ mod tests {
     }
 
     #[test]
-    fn infers_revoke_keyword() {
+    fn infers_grant_keyword() {
         let (pos, query) = with_pos(format!(
             r#"
-            revoke {}
+            grant {}
         "#,
             CURSOR_POS
         ));
 
-        let context = RevokeParser::get_context(query.as_str(), pos);
+        let context = GrantParser::get_context(query.as_str(), pos);
 
         assert_eq!(
             context,
-            RevokeContext {
+            GrantContext {
                 table_name: None,
                 schema_name: None,
                 node_text: "REPLACED_TOKEN".into(),
-                node_range: TextRange::new(TextSize::new(20), TextSize::new(34)),
-                node_kind: "revoke_role".into(),
+                node_range: TextRange::new(TextSize::new(19), TextSize::new(33)),
+                node_kind: "grant_role".into(),
             }
         );
     }
@@ -226,21 +234,44 @@ mod tests {
     fn infers_table_name() {
         let (pos, query) = with_pos(format!(
             r#"
-            revoke select on {} 
+            grant select on {} 
         "#,
             CURSOR_POS
         ));
 
-        let context = RevokeParser::get_context(query.as_str(), pos);
+        let context = GrantParser::get_context(query.as_str(), pos);
 
         assert_eq!(
             context,
-            RevokeContext {
+            GrantContext {
                 table_name: None,
                 schema_name: None,
                 node_text: "REPLACED_TOKEN".into(),
-                node_range: TextRange::new(TextSize::new(30), TextSize::new(44)),
-                node_kind: "revoke_table".into(),
+                node_range: TextRange::new(TextSize::new(29), TextSize::new(43)),
+                node_kind: "grant_table".into(),
+            }
+        );
+    }
+
+    #[test]
+    fn infers_table_name_with_keyword() {
+        let (pos, query) = with_pos(format!(
+            r#"
+            grant select on table {} 
+        "#,
+            CURSOR_POS
+        ));
+
+        let context = GrantParser::get_context(query.as_str(), pos);
+
+        assert_eq!(
+            context,
+            GrantContext {
+                table_name: None,
+                schema_name: None,
+                node_text: "REPLACED_TOKEN".into(),
+                node_range: TextRange::new(TextSize::new(35), TextSize::new(49)),
+                node_kind: "grant_table".into(),
             }
         );
     }
@@ -249,21 +280,44 @@ mod tests {
     fn infers_schema_and_table_name() {
         let (pos, query) = with_pos(format!(
             r#"
-            revoke select on public.{} 
+            grant select on public.{} 
         "#,
             CURSOR_POS
         ));
 
-        let context = RevokeParser::get_context(query.as_str(), pos);
+        let context = GrantParser::get_context(query.as_str(), pos);
 
         assert_eq!(
             context,
-            RevokeContext {
+            GrantContext {
                 table_name: None,
                 schema_name: Some("public".into()),
                 node_text: "REPLACED_TOKEN".into(),
-                node_range: TextRange::new(TextSize::new(37), TextSize::new(51)),
-                node_kind: "revoke_table".into(),
+                node_range: TextRange::new(TextSize::new(36), TextSize::new(50)),
+                node_kind: "grant_table".into(),
+            }
+        );
+    }
+
+    #[test]
+    fn infers_schema_and_table_name_with_keyword() {
+        let (pos, query) = with_pos(format!(
+            r#"
+            grant select on table public.{} 
+        "#,
+            CURSOR_POS
+        ));
+
+        let context = GrantParser::get_context(query.as_str(), pos);
+
+        assert_eq!(
+            context,
+            GrantContext {
+                table_name: None,
+                schema_name: Some("public".into()),
+                node_text: "REPLACED_TOKEN".into(),
+                node_range: TextRange::new(TextSize::new(42), TextSize::new(56)),
+                node_kind: "grant_table".into(),
             }
         );
     }
@@ -272,44 +326,44 @@ mod tests {
     fn infers_role_name() {
         let (pos, query) = with_pos(format!(
             r#"
-            revoke select on public.users from {} 
+            grant select on public.users to {} 
         "#,
             CURSOR_POS
         ));
 
-        let context = RevokeParser::get_context(query.as_str(), pos);
+        let context = GrantParser::get_context(query.as_str(), pos);
 
         assert_eq!(
             context,
-            RevokeContext {
+            GrantContext {
                 table_name: Some("users".into()),
                 schema_name: Some("public".into()),
                 node_text: "REPLACED_TOKEN".into(),
-                node_range: TextRange::new(TextSize::new(48), TextSize::new(62)),
-                node_kind: "revoke_role".into(),
+                node_range: TextRange::new(TextSize::new(45), TextSize::new(59)),
+                node_kind: "grant_role".into(),
             }
         );
     }
 
     #[test]
-    fn infers_multiple_roles() {
+    fn determines_table_name_after_schema() {
         let (pos, query) = with_pos(format!(
             r#"
-            revoke select on public.users from alice, {}
+            grant select on public.{} to test_role
         "#,
             CURSOR_POS
         ));
 
-        let context = RevokeParser::get_context(query.as_str(), pos);
+        let context = GrantParser::get_context(query.as_str(), pos);
 
         assert_eq!(
             context,
-            RevokeContext {
-                table_name: Some("users".into()),
+            GrantContext {
+                table_name: None,
                 schema_name: Some("public".into()),
                 node_text: "REPLACED_TOKEN".into(),
-                node_range: TextRange::new(TextSize::new(55), TextSize::new(69)),
-                node_kind: "revoke_role".into(),
+                node_range: TextRange::new(TextSize::new(36), TextSize::new(50)),
+                node_kind: "grant_table".into(),
             }
         );
     }
@@ -318,21 +372,44 @@ mod tests {
     fn infers_quoted_schema_and_table() {
         let (pos, query) = with_pos(format!(
             r#"
-            revoke select on "MySchema"."MyTable" from {}
+            grant select on "MySchema"."MyTable" to {}
         "#,
             CURSOR_POS
         ));
 
-        let context = RevokeParser::get_context(query.as_str(), pos);
+        let context = GrantParser::get_context(query.as_str(), pos);
 
         assert_eq!(
             context,
-            RevokeContext {
+            GrantContext {
                 table_name: Some("MyTable".into()),
                 schema_name: Some("MySchema".into()),
                 node_text: "REPLACED_TOKEN".into(),
-                node_range: TextRange::new(TextSize::new(56), TextSize::new(70)),
-                node_kind: "revoke_role".into(),
+                node_range: TextRange::new(TextSize::new(53), TextSize::new(67)),
+                node_kind: "grant_role".into(),
+            }
+        );
+    }
+
+    #[test]
+    fn infers_multiple_roles() {
+        let (pos, query) = with_pos(format!(
+            r#"
+            grant select on public.users to alice, {}
+        "#,
+            CURSOR_POS
+        ));
+
+        let context = GrantParser::get_context(query.as_str(), pos);
+
+        assert_eq!(
+            context,
+            GrantContext {
+                table_name: Some("users".into()),
+                schema_name: Some("public".into()),
+                node_text: "REPLACED_TOKEN".into(),
+                node_range: TextRange::new(TextSize::new(52), TextSize::new(66)),
+                node_kind: "grant_role".into(),
             }
         );
     }
