+++ crates/pgt_treesitter/src/queries/relations.rs
@@ -1,6 +1,6 @@
 use std::sync::LazyLock;
 
-use crate::{Query, QueryResult};
+use crate::queries::{Query, QueryResult};
 
 use super::QueryTryFrom;
 
@@ -79,7 +79,7 @@ impl<'a> QueryTryFrom<'a> for RelationMatch<'a> {
 }
 
 impl<'a> Query<'a> for RelationMatch<'a> {
-    fn execute(root_node: tree_sitter::Node<'a>, stmt: &'a str) -> Vec<crate::QueryResult<'a>> {
+    fn execute(root_node: tree_sitter::Node<'a>, stmt: &'a str) -> Vec<QueryResult<'a>> {
         let mut cursor = tree_sitter::QueryCursor::new();
 
         let matches = cursor.matches(&TS_QUERY, root_node, stmt.as_bytes());
@@ -112,8 +112,9 @@ impl<'a> Query<'a> for RelationMatch<'a> {
 
 #[cfg(test)]
 mod tests {
+    use crate::queries::TreeSitterQueriesExecutor;
+
     use super::RelationMatch;
-    use crate::TreeSitterQueriesExecutor;
 
     #[test]
     fn finds_table_without_schema() {
