+++ crates/pgt_treesitter/src/queries/select_columns.rs
@@ -1,6 +1,6 @@
 use std::sync::LazyLock;
 
-use crate::{Query, QueryResult};
+use crate::queries::{Query, QueryResult};
 
 use super::QueryTryFrom;
 
@@ -63,7 +63,7 @@ impl<'a> QueryTryFrom<'a> for SelectColumnMatch<'a> {
 }
 
 impl<'a> Query<'a> for SelectColumnMatch<'a> {
-    fn execute(root_node: tree_sitter::Node<'a>, stmt: &'a str) -> Vec<crate::QueryResult<'a>> {
+    fn execute(root_node: tree_sitter::Node<'a>, stmt: &'a str) -> Vec<QueryResult<'a>> {
         let mut cursor = tree_sitter::QueryCursor::new();
 
         let matches = cursor.matches(&TS_QUERY, root_node, stmt.as_bytes());
@@ -96,7 +96,7 @@ impl<'a> Query<'a> for SelectColumnMatch<'a> {
 
 #[cfg(test)]
 mod tests {
-    use crate::TreeSitterQueriesExecutor;
+    use crate::queries::TreeSitterQueriesExecutor;
 
     use super::SelectColumnMatch;
